fix(config): merge dead next.config.ts into next.config.js — restore prod security headers
next.config.ts existed but had no `export default`, leaving its features silently
dead in production: removeConsole console-stripping, X-Frame-Options/X-Content-Type-Options/
X-XSS-Protection/Referrer-Policy on /api/*, Cache-Control: no-store on /api/cron/*,
and DEPLOYMENT_PLATFORM env. All merged into the canonical next.config.js. Documented
in CLAUDE.md (item #18).

Subtask 2 commit 1 of dashboard-streaming-perf-v2 plan — unblocks the staleTimes
config change in the next commit.
