--- next.config.ts
@@ -1,56 +1,0 @@
-import type { NextConfig } from "next";
-
-const nextConfig: NextConfig = {
-  // Production optimization for Railway deployment
-  // Note: Edge Runtime should be configured per-route, not globally
-  
-  // Production optimization for Railway
-  compiler: {
-    // Remove console.log in production builds
-    removeConsole: process.env.NODE_ENV === 'production' ? {
-      exclude: ['error', 'warn'],
-    } : false,
-  },
-  
-  // Environment-specific configuration
-  env: {
-    // Platform detection for monitoring
-    DEPLOYMENT_PLATFORM: 'VERCEL',
-  },
-  
-  // Headers for security and performance
-  async headers() {
-    return [
-      {
-        source: '/api/(.*)',
-        headers: [
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          {
-            key: 'X-Frame-Options',
-            value: 'DENY',
-          },
-          {
-            key: 'X-XSS-Protection',
-            value: '1; mode=block',
-          },
-          {
-            key: 'Referrer-Policy',
-            value: 'strict-origin-when-cross-origin',
-          },
-        ],
-      },
-      {
-        source: '/api/cron/(.*)',
-        headers: [
-          {
-            key: 'Cache-Control',
-            value: 'no-store, no-cache, must-revalidate',
-          },
-        ],
-      },
-    ];
-  },
-};
